Move RelaySwitch to switch.py
